requirements.txt
pyyaml>=5.1
requests>=2.22.0
markdown>=3.1.1
weasyprint>=51.0
beautifulsoup4>=4.8.0